requirements.txt
openpyxl==3.1.1
paho_mqtt==1.6.1
requests==2.28.2
